Changed according to code reviews
